+++ src/frontend/apps/e2e/.env.example
@@ -2,7 +2,6 @@ PORT=3000
 BASE_URL=http://localhost:3000
 BASE_API_URL=http://localhost:8071/api/v1.0
 COLLABORATION_WS_URL=ws://localhost:4444/collaboration/ws/
-COLLABORATION_WS_NOT_CONNECTED_READY_ONLY=true
 MEDIA_BASE_URL=http://localhost:8083
 IS_INSTANCE=false
 CUSTOM_SIGN_IN=false
